using slf4j as logging abstraction
