+++ citydb-cli/build.gradle
@@ -19,7 +19,9 @@ dependencies {
     implementation project(':citydb-config')
     implementation project(':citydb-database-postgres')
     implementation project(':citydb-io-citygml')
-    implementation project(':citydb-logging')
+    implementation 'org.apache.logging.log4j:log4j-api:2.25.1'
+    implementation 'org.apache.logging.log4j:log4j-core:2.25.1'
+    implementation 'org.apache.logging.log4j:log4j-slf4j2-impl:2.25.1'
     citydb 'org.citydb:3dcitydb:5.1.0-rc.2'
 }
