+++ citydb-cli/src/main/java/org/citydb/cli/exporter/ExportController.java
@@ -29,6 +29,7 @@ import org.citydb.cli.exporter.options.QueryOptions;
 import org.citydb.cli.exporter.options.TilingOptions;
 import org.citydb.cli.exporter.util.SequentialWriter;
 import org.citydb.cli.exporter.util.TilingHelper;
+import org.citydb.cli.logging.LoggerManager;
 import org.citydb.cli.util.CommandHelper;
 import org.citydb.cli.util.FeatureStatistics;
 import org.citydb.config.Config;
@@ -45,7 +46,6 @@ import org.citydb.io.writer.FeatureWriter;
 import org.citydb.io.writer.WriteException;
 import org.citydb.io.writer.WriteOptions;
 import org.citydb.io.writer.options.OutputFormatOptions;
-import org.citydb.logging.LoggerManager;
 import org.citydb.model.feature.Feature;
 import org.citydb.operation.exporter.Exporter;
 import org.citydb.operation.exporter.options.AppearanceOptions;
