+++ citydb-cli/src/main/java/org/citydb/cli/importer/ImportController.java
@@ -29,6 +29,7 @@ import org.citydb.cli.importer.duplicate.DuplicateController;
 import org.citydb.cli.importer.filter.Filter;
 import org.citydb.cli.importer.options.FilterOptions;
 import org.citydb.cli.importer.options.MetadataOptions;
+import org.citydb.cli.logging.LoggerManager;
 import org.citydb.cli.util.CommandHelper;
 import org.citydb.config.Config;
 import org.citydb.config.ConfigException;
@@ -43,7 +44,6 @@ import org.citydb.io.reader.FeatureReader;
 import org.citydb.io.reader.ReadOptions;
 import org.citydb.io.reader.filter.FilterException;
 import org.citydb.io.reader.options.InputFormatOptions;
-import org.citydb.logging.LoggerManager;
 import org.citydb.model.feature.Feature;
 import org.citydb.operation.importer.Importer;
 import picocli.CommandLine;
