+++ citydb-cli/src/main/java/org/citydb/cli/importer/duplicate/DuplicateDeleter.java
@@ -25,8 +25,8 @@ import org.apache.logging.log4j.Logger;
 import org.citydb.cli.ExecutionException;
 import org.citydb.cli.importer.ImportMode;
 import org.citydb.cli.importer.ImportOptions;
+import org.citydb.cli.logging.LoggerManager;
 import org.citydb.database.adapter.DatabaseAdapter;
-import org.citydb.logging.LoggerManager;
 import org.citydb.operation.deleter.DeleteOptions;
 import org.citydb.operation.deleter.Deleter;
 import org.citydb.operation.deleter.options.DeleteMode;
