+++ citydb-cli/src/main/java/org/citydb/cli/info/InfoCommand.java
@@ -9,13 +9,13 @@ import org.citydb.cli.common.Command;
 import org.citydb.cli.common.ConfigOption;
 import org.citydb.cli.common.ConnectionOptions;
 import org.citydb.cli.common.ThreadsOptions;
+import org.citydb.cli.logging.LoggerManager;
 import org.citydb.cli.util.CommandHelper;
 import org.citydb.cli.util.Streams;
 import org.citydb.config.Config;
 import org.citydb.config.ConfigException;
 import org.citydb.database.DatabaseManager;
 import org.citydb.database.util.IndexHelper;
-import org.citydb.logging.LoggerManager;
 import org.citydb.util.report.DatabaseReport;
 import org.citydb.util.report.ReportOptions;
 import picocli.CommandLine;
