+++ citydb-cli/src/main/java/org/citydb/cli/util/FeatureStatistics.java
@@ -23,9 +23,9 @@ package org.citydb.cli.util;
 
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
+import org.citydb.cli.logging.LoggerManager;
 import org.citydb.database.adapter.DatabaseAdapter;
 import org.citydb.database.schema.SchemaMapping;
-import org.citydb.logging.LoggerManager;
 import org.citydb.model.appearance.Appearance;
 import org.citydb.model.common.Name;
 import org.citydb.model.common.Namespaces;
