+++ citydb-core/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module org.citydb.core {
     requires org.apache.commons.io;
     requires com.h2database.mvstore;
     requires transitive com.alibaba.fastjson2;
+    requires transitive org.slf4j;
     requires transitive org.apache.tika.core;
 
     exports org.citydb.core;
