+++ citydb-database/build.gradle
@@ -5,7 +5,6 @@ def geotools_version = '33.2'
 dependencies {
     api project(':citydb-core')
     api project(':citydb-config')
-    api project(':citydb-logging')
     api project(':citydb-model')
     api 'org.citydb:sqlbuilder:3.0.5'
     api "org.geotools:gt-referencing:$geotools_version"
