+++ citydb-database/src/main/java/org/citydb/database/DatabaseManager.java
@@ -21,8 +21,6 @@
 
 package org.citydb.database;
 
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.Logger;
 import org.apache.tomcat.jdbc.pool.DataSource;
 import org.apache.tomcat.jdbc.pool.PoolProperties;
 import org.citydb.database.adapter.DatabaseAdapter;
@@ -30,7 +28,6 @@ import org.citydb.database.adapter.DatabaseAdapterException;
 import org.citydb.database.adapter.DatabaseAdapterManager;
 import org.citydb.database.connection.ConnectionDetails;
 import org.citydb.database.connection.PoolOptions;
-import org.citydb.logging.LoggerManager;
 
 import java.sql.SQLException;
 import java.util.Objects;
@@ -38,7 +35,6 @@ import java.util.Properties;
 import java.util.function.Consumer;
 
 public class DatabaseManager {
-    private final Logger logger = LoggerManager.getInstance().getLogger(DatabaseManager.class);
     private DatabaseAdapter adapter;
     private DataSource dataSource;
 
@@ -122,11 +118,7 @@ public class DatabaseManager {
         return dataSource;
     }
 
-    public void logDatabaseMetadata(Level level) {
-        printDatabaseMetadata(s -> logger.log(level, s));
-    }
-
-    public void printDatabaseMetadata(Consumer<String> consumer) {
+    public void reportDatabaseMetadata(Consumer<String> consumer) {
         if (isConnected()) {
             consumer.accept("3D City Database: " + adapter.getDatabaseMetadata().getVersion());
             consumer.accept("DBMS: " + adapter.getDatabaseMetadata().getVendorProductString());
