+++ citydb-io-citygml/build.gradle
@@ -1,7 +1,6 @@
 def citygml4j_version = '3.2.4'
 
 dependencies {
-    api project(':citydb-logging')
     api "org.citygml4j:citygml4j-xml:$citygml4j_version"
     api "org.citygml4j:citygml4j-cityjson:$citygml4j_version"
     implementation project(':citydb-io')
