+++ citydb-io-citygml/src/main/java/org/citydb/io/citygml/adapter/geometry/builder/GeometryHelper.java
@@ -21,7 +21,6 @@
 
 package org.citydb.io.citygml.adapter.geometry.builder;
 
-import org.apache.logging.log4j.Level;
 import org.citydb.io.citygml.adapter.appearance.builder.AppearanceHelper;
 import org.citydb.io.citygml.builder.ModelBuildException;
 import org.citydb.io.citygml.reader.ModelBuilderHelper;
@@ -31,6 +30,7 @@ import org.citydb.model.common.Reference;
 import org.citydb.model.geometry.Geometry;
 import org.citydb.model.geometry.ImplicitGeometry;
 import org.citydb.model.property.ImplicitGeometryProperty;
+import org.slf4j.event.Level;
 import org.xmlobjects.gml.model.geometry.AbstractGeometry;
 import org.xmlobjects.gml.model.geometry.GeometryProperty;
