+++ citydb-io-citygml/src/main/java/org/citydb/io/citygml/writer/CityJSONWriter.java
@@ -21,7 +21,6 @@
 
 package org.citydb.io.citygml.writer;
 
-import org.apache.logging.log4j.Logger;
 import org.citydb.config.ConfigException;
 import org.citydb.core.cache.PersistentMapStore;
 import org.citydb.core.concurrent.CountLatch;
@@ -32,7 +31,6 @@ import org.citydb.io.citygml.writer.util.GlobalFeatureWriter;
 import org.citydb.io.writer.FeatureWriter;
 import org.citydb.io.writer.WriteException;
 import org.citydb.io.writer.WriteOptions;
-import org.citydb.logging.LoggerManager;
 import org.citydb.model.feature.Feature;
 import org.citydb.model.geometry.ImplicitGeometry;
 import org.citygml4j.cityjson.CityJSONContext;
@@ -40,6 +38,8 @@ import org.citygml4j.cityjson.writer.AbstractCityJSONWriter;
 import org.citygml4j.core.model.appearance.Appearance;
 import org.citygml4j.core.model.core.AbstractAppearanceProperty;
 import org.citygml4j.core.model.core.AbstractFeature;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.xmlobjects.gml.model.geometry.AbstractGeometry;
 
 import java.io.IOException;
@@ -48,7 +48,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 
 public class CityJSONWriter implements FeatureWriter, GlobalFeatureWriter {
-    private final Logger logger = LoggerManager.getInstance().getLogger(CityJSONWriter.class);
+    private final Logger logger = LoggerFactory.getLogger(CityJSONWriter.class);
     private final AbstractCityJSONWriter<?> writer;
     private final PersistentMapStore store;
     private final ExecutorService service;
