--- citydb-logging/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-    api 'org.apache.logging.log4j:log4j-api:2.25.1'
-    implementation 'org.apache.logging.log4j:log4j-core:2.25.1'
-    implementation 'org.slf4j:slf4j-nop:2.0.17'
-}
