+++ settings.gradle
@@ -9,7 +9,6 @@ include 'citydb-database'
 include 'citydb-database-postgres'
 include 'citydb-io'
 include 'citydb-io-citygml'
-include 'citydb-logging'
 include 'citydb-model'
 include 'citydb-operation'
 include 'citydb-plugin'
